Don't save empty caption region window if a real one was set before it
only real changes here are lines 147 148 and 289 290. The initial capture region window will be empty ('') but when a real window is set, the save setting for the capture region will only be a real window. This is really useful for when you, for example, close OBS before AutoSplit. It will save the OBS window setting instead of the empty window setting.