Drop PyQt6 in favor of PySide6 (#204)
We already have to install PySide for its tooling anyway. With this we would use the official bindings with a public bug tracker. The size of the bundle is also reduced by ~10MB.

I was also able to remove a hack for the shortcutContext breaking the UI generation tool
Unblocks #209