implemented text recognition (ocr) (#272)
This could be a useful addition in cases where modern game rendering
and visual effects (clutter) cause difficulties to find good comparison
images.

It currently depends on pytesseract and Tesseract-OCR but tests
with EasyOCR have also been conducted. Both seem to get similar good
recognition results. EasyOCR looks like to cause higher CPU load then
tesseract. Tesseract on the other hand is an external dependency that
needs to be installed seperatly.

The text comparison of the expected and recognized string has two modes.
A the levenshtein ratio or a partial string match

* display text that is searched for
* call tesseract ourselfs which ditches all python binding libraries to not include Pillow
* include PATH variable use

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Avasam <samuel.06@hotmail.com>